requirements.txt
Flask==2.2.3
openai==0.27.2
python-dotenv==1.0.0
requests==2.28.1
